+++ CHANGES.rst
@@ -18,6 +18,7 @@ Unreleased
 * :meth:`~Reader.get_feeds` now sorts feeds in a case insensitive way. (`#103`_)
 * Add ``order_by`` argument to :meth:`~Reader.get_feeds`, allowing to sort
   feeds by title or by when they were added. (`#98`_)
+* Allow changing the feed sort order in the web application. (`#98`_)
 
 .. _#99: https://github.com/lemon24/reader/issues/99
 .. _#102: https://github.com/lemon24/reader/issues/102